+++ modules/apps/layout/layout-locked-layouts-web/src/main/java/com/liferay/layout/locked/layouts/web/internal/configuration/admin/display/LockedLayoutsSiteSettingsConfigurationScreenWrapper.java
@@ -3,13 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.locked.layouts.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.layout.locked.layouts.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
-import com.liferay.layout.locked.layouts.web.internal.configuration.LockedLayoutsCompanyConfiguration;
+import com.liferay.layout.configuration.LockedLayoutsGroupConfiguration;
 import com.liferay.layout.locked.layouts.web.internal.display.context.LockedLayoutsConfigurationDisplayContext;
-import com.liferay.layout.locked.layouts.web.internal.display.context.LockedLayoutsPortalSettingsConfigurationDisplayContext;
+import com.liferay.layout.locked.layouts.web.internal.display.context.LockedLayoutsSiteSettingsConfigurationDisplayContext;
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
@@ -20,8 +20,8 @@ import com.liferay.portal.kernel.module.configuration.ConfigurationException;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenFactory;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
 import java.io.IOException;
 
@@ -38,16 +38,16 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Lourdes Fernández Besada
+ * @author Mikel Lorza
  */
 @Component(service = ConfigurationScreen.class)
-public class LockedLayoutsPortalSettingsConfigurationScreenWrapper
+public class LockedLayoutsSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
-		return _portalSettingsConfigurationScreenFactory.create(
-			new LockedLayoutsPortalSettingsConfigurationScreenContributor());
+		return _siteSettingsConfigurationScreenFactory.create(
+			new LockedLayoutsSiteSettingsConfigurationScreenContributor());
 	}
 
 	@Reference
@@ -59,17 +59,17 @@ public class LockedLayoutsPortalSettingsConfigurationScreenWrapper
 	@Reference
 	private Language _language;
 
-	@Reference
-	private PortalSettingsConfigurationScreenFactory
-		_portalSettingsConfigurationScreenFactory;
-
 	@Reference(
 		target = "(osgi.web.symbolicname=com.liferay.layout.locked.layouts.web)"
 	)
 	private ServletContext _servletContext;
 
-	private class LockedLayoutsPortalSettingsConfigurationScreenContributor
-		implements PortalSettingsConfigurationScreenContributor {
+	@Reference
+	private SiteSettingsConfigurationScreenFactory
+		_siteSettingsConfigurationScreenFactory;
+
+	private class LockedLayoutsSiteSettingsConfigurationScreenContributor
+		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
@@ -83,7 +83,7 @@ public class LockedLayoutsPortalSettingsConfigurationScreenWrapper
 
 		@Override
 		public String getKey() {
-			return "locked-layouts-portal-settings";
+			return "site-configuration-locked-layouts";
 		}
 
 		@Override
@@ -93,7 +93,7 @@ public class LockedLayoutsPortalSettingsConfigurationScreenWrapper
 
 		@Override
 		public String getSaveMVCActionCommandName() {
-			return "/instance_settings/save_locked_layouts_instance_settings";
+			return "/site_settings/save_locked_layouts_site_settings";
 		}
 
 		@Override
@@ -113,26 +113,26 @@ public class LockedLayoutsPortalSettingsConfigurationScreenWrapper
 			try {
 				httpServletRequest.setAttribute(
 					LockedLayoutsConfigurationDisplayContext.class.getName(),
-					new LockedLayoutsPortalSettingsConfigurationDisplayContext(
-						_hasConfiguration(themeDisplay.getCompanyId()),
-						_configurationProvider.getCompanyConfiguration(
-							LockedLayoutsCompanyConfiguration.class,
-							themeDisplay.getCompanyId())));
+					new LockedLayoutsSiteSettingsConfigurationDisplayContext(
+						_hasConfiguration(themeDisplay.getScopeGroupId()),
+						_configurationProvider.getGroupConfiguration(
+							LockedLayoutsGroupConfiguration.class,
+							themeDisplay.getScopeGroupId())));
 			}
 			catch (PortalException portalException) {
 				ReflectionUtil.throwException(portalException);
 			}
 		}
 
-		private boolean _hasConfiguration(long companyId)
+		private boolean _hasConfiguration(long groupId)
 			throws ConfigurationException {
 
 			try {
 				String filterString = StringBundler.concat(
 					"(&(", ConfigurationAdmin.SERVICE_FACTORYPID,
 					StringPool.EQUAL,
-					LockedLayoutsCompanyConfiguration.class.getName(),
-					".scoped)(companyId=", companyId, "))");
+					LockedLayoutsGroupConfiguration.class.getName(), ".scoped",
+					")(groupId=", groupId, "))");
 
 				Configuration[] configuration =
 					_configurationAdmin.listConfigurations(filterString);
